Batch: pack DPIC to reduce sync times
We pack DPIC of DifftestBundles on Hardware and unpack it in software.
So tens of DPIC called times will be reduced to one, which reduce
sync times in platform such as palladium.

Difftest cycle delayed for DifftestBundles to collect valid bundles
in bundleNum cycles, which cause delayed Comparision result.

BatchInterval and BatchFinish are appended to infoVec to seperate DPIC
packed for different Cycle. Other elements contains BundleType and
BundleLen to help unpack DataVec.

Now we set maximum byte packed as 4096.